refactor: notebook view, extract components
