+++ apps/ui/components/notebook/MonacoEditorClient.tsx
@@ -1,0 +1,13 @@
+"use client";
+
+import dynamic from "next/dynamic";
+
+const MonacoEditor = dynamic(
+  async () => {
+    const mod = await import("@monaco-editor/react");
+    return mod.default;
+  },
+  { ssr: false }
+);
+
+export default MonacoEditor;
